+++ routes/userRoutes/__init__.py
@@ -1,2 +1,1 @@
 from .users import *
-from .authorization import *
